refactor: 데이터와 연관된 부분 model로 분리
